Moves SessionNew to button modal in SeasonsIndex
