+++ src/Content.jsx
@@ -1,7 +1,6 @@
 import axios from "axios";
 import { useState, useEffect } from "react";
 import { SessionsIndex } from "./SessionsIndex";
-import { SessionsNew } from "./SessionsNew";
 import { SessionsShow } from "./SessionsShow";
 import { Modal } from "./Modal";
 import { SeasonsIndex } from "./SeasonsIndex";
@@ -44,14 +43,6 @@ export function Content() {
     axios.get("http://localhost:3000/league_sessions.json").then((response) => {
       console.log(response.data);
       setSessions(response.data);
-    });
-  };
-
-  const handleCreateSession = (params, successCallback) => {
-    console.log("handleCreateSession", params);
-    axios.post("http://localhost:3000/league_sessions.json", params).then((response) => {
-      setSessions([...sessions, response.data]);
-      successCallback();
     });
   };
 
@@ -108,7 +99,6 @@ export function Content() {
         <SeasonsNew onCreateSeason={handleCreateSeason} />
       </Modal>
       <SeasonsIndex seasons={seasons} />
-      <SessionsNew onCreateSession={handleCreateSession} />
       <SessionsIndex sessions={sessions} onShowSession={handleShowSession} />
       <Modal show={isSessionsShowVisible} onClose={handleClose}>
         <SessionsShow
